modular_triager/requirements.txt
streamlit
boto3
pandas
requests
streamlit-extras
strands-agents
strands-agents-tools


